enhance upload from file post request to handle mp3

+++ backend/cmd/clear.py
@@ -4,6 +4,7 @@ upload_path = Path("uploads")
 
 def clear_uploads_dir():
   for file in upload_path.iterdir():
-    file.unlink()
+    if file.suffix in {".mp3", ".mp4"}:
+      file.unlink()
 
 clear_uploads_dir()
